# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para operaciones del sistema SistaL.",
        "title": "MS Operations API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8083",
    "basePath": "/api/v1",
    "paths": {
...
        "/peticiones": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene lista paginada de peticiones de uniforme con filtros opcionales",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peticiones"
                ],
                "summary": "Obtener peticiones",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Número de página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Elementos por página",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filtrar por ID de funcionario",
                        "name": "id_funcionario",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filtrar por ID de estado",
                        "name": "id_estado",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filtrar por ID de tipo petición",
                        "name": "id_tipo_peticion",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filtrar por ID de temporada",
                        "name": "id_temporada",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por maternal (true/false)",
                        "name": "maternal",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar con/sin despacho (true/false)",
                        "name": "con_despacho",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha desde (YYYY-MM-DD)",
                        "name": "fecha_desde",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha hasta (YYYY-MM-DD)",
                        "name": "fecha_hasta",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Búsqueda en observaciones",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Campo para ordenar (default fecha_registro)",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Orden (asc/desc, default desc)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de peticiones con paginación",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea una nueva petición de uniforme con tallajes asociados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peticiones"
                ],
                "summary": "Crear petición",
                "parameters": [
                    {
                        "description": "Datos de la petición",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreatePeticionDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Petición creada exitosamente",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Datos de entrada inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/peticiones/{id_peticion}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene detalles completos de una petición específica incluyendo tallajes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peticiones"
                ],
                "summary": "Obtener petición por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la petición",
                        "name": "id_peticion",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la petición",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Petición no encontrada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Actualiza campos de una petición existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peticiones"
                ],
                "summary": "Actualizar petición",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la petición",
                        "name": "id_peticion",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos a actualizar",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdatePeticionDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Petición actualizada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Datos de entrada inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Petición no encontrada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/peticiones/{id_peticion}/estado": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cambia el estado de una petición con validación de transiciones",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peticiones"
                ],
                "summary": "Cambiar estado de petición",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la petición",
                        "name": "id_peticion",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nuevo estado",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CambiarEstadoPeticionDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado actualizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Transición de estado inválida",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Petición no encontrada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CambiarEstadoPeticionDTO": {
            "type": "object",
            "required": [
                "id_estado"
            ],
            "properties": {
                "id_estado": {
                    "type": "integer"
                },
                "observacion": {
                    "type": "string"
                }
            }
        },
...
        "handlers.CreatePeticionDTO": {
            "type": "object",
            "required": [
                "id_funcionario",
                "id_tipo_peticion",
                "id_uniforme",
                "tallaje"
            ],
            "properties": {
                "id_estado": {
                    "type": "integer"
                },
                "id_funcionario": {
                    "type": "integer"
                },
                "id_temporada": {
                    "type": "integer"
                },
                "id_tipo_peticion": {
                    "type": "integer"
                },
                "id_uniforme": {
                    "type": "integer"
                },
                "maternal": {
                    "type": "boolean"
                },
                "observacion": {
                    "type": "string"
                },
                "tallaje": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/handlers.TallajeItemDTO"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Token de autenticación JWT en formato Bearer",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}